==> notebooks/01_PREPARE_SAMPLE_GPS.ipynb <==
# About
Given notebook is created to prepare October GPS data for clustering

import os
import numpy as np
import pandas as pd
import pymongo
from sklearn.cluster import DBSCAN
import plotly.express as px
import geopandas as gpd


ATLAS_URL="mongodb+srv://waste_labs_admin:WaSte_LaBs_password!1@dev-instance.f4t0c.mongodb.net/waste_labs?retryWrites=true&w=majority"
ATLAS_DATABASE='waste_labs'
ATLAS_GPS_COLLECTION='GPS_RECORDS'
ATLAS_STATUS_COLLECTION='GPS_RECORDS_LOGS'

S_OCTOBER_GPS_DIR = '/Data/Source/OctoberGPS/'
S_GPS_DIR = '/Data/Source/migrator_outputs/'

I_MAP_MATCHING_REPORT = '/Data/Intermediate/MapMatchingReports/MapMatching_october2020.csv'

!ls /Data/Source/

# Data

### 1. Map matching logs from mongoDB

def get_map_matching_logs():
    with pymongo.MongoClient(ATLAS_URL) as client:
        db = client[ATLAS_DATABASE]
        collection = db[ATLAS_STATUS_COLLECTION]
        iterator = collection.aggregate([
            {'$addFields': {
                'match_API_URL': '$matching_parameters.API_URL',
                'additional_match_params': '$matching_parameters.additional_parameters',
                'radius_of_match': '$matching_parameters.radius_of_error',
                'n_processes_match': '$matching_parameters.n_processes',
                'n_points_match': '$matching_metrics.n_points_mat',
                'n_points_match_perc': '$matching_metrics.n_points_mat_perc',
                'matching_time_in_sec': '$matching_metrics.matching_time'
            }},
            {'$project': {'_id': 0,
                          'matching_parameters': 0,
                          'matching_metrics': 0
                         }},
        ])
        logs = pd.DataFrame(list(iterator))

    return logs

logs = get_map_matching_logs()

# Postprocessing
logs = logs.loc[logs.status == 'succesfull', ]
logs = logs.loc[pd.to_datetime(logs.waste_collection_date).dt.month == 10, :]
logs['csv_file'] = logs.filename.str.replace('.gpx', '.csv')
logs.to_csv(I_MAP_MATCHING_REPORT, index=False)

# 2. Migration

# 2. building source and target path
fnames = logs.filename.str.replace('.gpx', '.csv')
source_paths = list(map(lambda x, y: x + y, [S_GPS_DIR]*logs.shape[0], fnames))
target_paths = list(map(lambda x, y: x + y, [S_OCTOBER_GPS_DIR]*logs.shape[0], fnames))

# 3. Migration
def migrate(src, trgt):
    pd.read_csv(src).to_csv(trgt, index=False)

%%time
from multiprocessing import Pool


with Pool(5) as pool:
    pool.starmap(migrate, zip(source_paths, target_paths))
