==> notebooks/05_VEHICLE_TIME_SHIFTS_DETECTION.ipynb <==
# About
In notebook 04, we've found that time shifts are different for different type of trucks.
So, in this notebook i'll try to detect time shifts for different trucks and routes

import os
import numpy as np
import pandas as pd
import logging
from shapely.geometry import Point
import plotly.express as px
import geopandas as gpd
import plotly.express as px
from multiprocessing import Pool
from scipy.stats import zscore
from GPSOdyssey import Polaris, Kepler, Void, Vega
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = 'all'

## Time shifts

* RFID: by default client provides time shifts in +08:00
* GPS: Different for different trucks

def generate_trip_id(gps):
    gps['trip_id'] = gps['date'].astype('str') + ' ' + gps['vehicle_id']
    return gps

def generate_point_id(gps):
    gps['point_id'] = gps['datetime'].astype('str') + ' ' + gps['vehicle_id']
    return gps

## Data

### Map matching report

map_matching_report = pd.read_csv('/Data/Intermediate/MapMatchingReports/MapMatching_october2020.csv')

map_matching_report = Polaris(map_matching_report) \
    .pandas_df_operation(func_name='rename', arguments={'columns': {'truck_id': 'vehicle_id'}}).df

map_matching_report['vehicle_id'] = map_matching_report['vehicle_id'].str.replace(' ', '-')

### 1. Route vehicle assignment

I_ROUTE_VEHICLE_ASSIGNMENT = '/Data/Intermediate/assignments/route_vehicle_assignments_october.csv'

route_vehic_assign = pd.read_csv(I_ROUTE_VEHICLE_ASSIGNMENT)
route_vehic_assign = route_vehic_assign[route_vehic_assign['route_type'] == 'JS-SE'] \
    .groupby(['route_type']).agg({'vehicle_id': 'unique'}).reset_index()

route_vehic_assign = route_vehic_assign.explode('vehicle_id')

### 2. RFID

S_RFID_OCTOBER = '/Data/Source/RFID/RFID_october.csv'


rfid = pd.read_csv(S_RFID_OCTOBER)
rfid['is_rfid_active'] = True

RFID_COLUMNS = ['date', 'time', 'unixtime', 'vehicle_id',
                'RFID', 'is_rfid_active', 'Latitude', 'Longitude']
rfid = rfid.loc[:, RFID_COLUMNS]
rfid.columns = [str(x).lower() for x in rfid.columns]

rfid = rfid.sort_values(by='unixtime')
rfid['datetime'] = rfid['date'] + ' ' + rfid['time']


rfid = generate_trip_id(rfid)
rfid = generate_point_id(rfid)

#### Select RFID day with larges amount records

# 1. Select only JS-SE + Count amount of records
rfid_agg = rfid.groupby(['date', 'vehicle_id']).agg(count=('time', 'count')).reset_index()

# 2. Select trucks of JS-SE route with largest amount of gps records
rfid_agg = rfid_agg.merge(rfid_agg.groupby(['vehicle_id']).agg(count=('count', 'max')).reset_index(),
                          how='inner', on=['vehicle_id', 'count'])

# 3. Rmoeve rec vehicle id
rfid_agg = rfid_agg[rfid_agg['vehicle_id'].str.slice(stop=2) == 'XE']

# 3. Select actual rfid
rfid = rfid.merge(rfid_agg[['date', 'vehicle_id']],
                  how='inner', on=['date', 'vehicle_id'])

### 3. GPS

map_matching_report = map_matching_report.merge(rfid_agg.rename(columns={'date': 'waste_collection_date'}), how='inner',
                                                on=['waste_collection_date', 'vehicle_id'])



%%time
S_CLUSTERED_GPS = '/Data/Source/OctoberGPS/'


vega = Vega(engine='pandas')
gps = vega.read_from_dir(directory=S_CLUSTERED_GPS,
                         file_extensions='.csv',
                         args={'parse_dates': ['time']},
                         concatenate=True, amt_in_parallel=6,
                         filenames=map_matching_report['csv_file'])


gps = Polaris(gps) \
    .select_columns(['truck_id', 'lon', 'lat', 'time', 'lon_match', 'lat_match']) \
    .pandas_df_operation(func_name='rename', arguments={'columns': {'truck_id': 'vehicle_id', 'time': 'datetime'}}) \
    .construct_datetime(datetime='datetime', offset='-02:00') \
    .remove_tz(time_col='datetime') \
    .add_date_col('datetime', 'date') \
    .add_time_col('datetime', 'time') \
    .add_unixtime('datetime', 'unixtime') \
    .df

gps = generate_trip_id(gps)
gps = generate_point_id(gps)
gps['unixtime'] = gps['unixtime'].astype('int')

# 4. RFID -> GPS

gps['datetime'] = gps['datetime'].astype('str')
rfid['datetime'] = rfid['datetime'].astype('str')

gps_m = gps.merge(rfid[['rfid', 'vehicle_id', 'datetime', 'unixtime',
                        'is_rfid_active', 'latitude', 'longitude']],
                  on=['vehicle_id', 'datetime'], how='outer')

gps_m['is_overlap'] = gps_m['lat_match'].isna()
gps_m.sort_values(['vehicle_id', 'datetime'], inplace=True)

cols2fill = ['lon', 'lat', 'time', 'vehicle_id', 'date', 'datetime', 'time',
             'unixtime_x', 'lon_match', 'lat_match', 'trip_id', 'point_id']
gps_m[cols2fill] = gps_m[cols2fill].fillna(method='ffill')

gps_m['is_rfid_active'] = ~gps_m['rfid'].isna()

gps_m[['latitude', 'longitude']] = gps_m[['latitude', 'longitude']].astype('float', errors='ignore')

# 4. Visualization

MAP_CONF_PATH = 'MAP_CONFIGS/05_TIME_SHIFT_ANALYSIS.json'

kepler = Kepler(data={'gps': gps_m, 'rfid': rfid}, height=800, config_path=MAP_CONF_PATH)
kepler.render_kepler_map()
kepler.get_rendered_map()

# kepler.save_map_config(path=MAP_CONF_PATH, overwrite_config=True)

# kepler.save_map_to_html_local('time_shifts_of_same_vehicle.html', overwrite_html=True)

# 5. Time shifts

RFID - GPS

route_vehic_assign

{
    'XE-5638S': '-02:00',
    'XE-5598Z': '-02:00',
    'XE-5628X': '-02:00',
    'XE-5577J': '-02:00',
    'XE-5665M': '-02:00',
    'XE-5629T': '-02:00',
    'XE-5632H': '-02:00',
    'XE-5680T': '-02:00',
    'XE-5559L': '-02:00',
    'XE-5610X': '-02:00',
    'XE-5636Y': '-02:00',
    'XE-5640J': '-02:00',
    'XE-5612R': '-02:00',
    # 'XE-5705G': '-02:00',
    'XE-5609B': '-01:00',
    'XE-5620S': '-01:00',
    'XE-5639P': '-01:00',
    'XE-5748T': '-01:00',
    'XE-5630M': '-01:00',
    # 'XE-5705G': '-01:00',
}

## Summary

At the end we see different time shifts for the same truck at different day.
